Tiger beetle integration

+++ backend/RetailBank/Models/CreateAccountRequest.cs
@@ -1,3 +1,3 @@
 ﻿namespace RetailBank.Models;
 
-public record CreateAccountRequest(UInt128 SalaryCents);
+public record CreateAccountRequest(UInt64 SalaryCents);
